refactor(clash): split the TS engine into orchestrator + pluggable kernel
Extract selection / exclusions / severity / identity / dedup / summary into a
backend-agnostic orchestrator (`runClash`) driven by a `ClashKernel` interface.
`TsKernel` wraps the existing BVH broad phase + exact triangle narrow phase as
the reference (and differential-test oracle); the Rust/WASM kernel will drop in
behind the same seam. Behaviour is unchanged — full suite still 72/72.
